refactor: optimize AstBuilderBase
